throw AtPublicKeyNotFound when current atsign public key is not found. Added unit tests
